+++ tests/unit/domain-step-form.spec.ts
@@ -26,7 +26,7 @@ describe('Domain Step Form', () => {
   runner.testCancel();
 
   it('should instantiate an autocomplete widget with proper options from the store', () => {
-    const wrapper = runner.shallowMount({ domains: ['foo', 'bar'] });
+    const wrapper = runner.shallowMount({ sources: ['foo', 'bar'] });
     const widgetAutocomplete = wrapper.find('autocompletewidget-stub');
 
     expect(widgetAutocomplete.attributes('options')).toEqual('foo,bar');